create visualization file
